Factor out cache-related structures from wave and move boo to kernels/boo.
Still need to modify wave to not rely on old cache manager.
Next steps:

Implement BooLaunchable for either Eager or Test contexts and work out some examples.
