--- iree/turbine/boo/bag/namespaces/auto_op/__init__.py
@@ -1,1 +1,0 @@
-from defs import *
