--- iree/turbine/boo/bag/namespaces/auto_op/defs.py
@@ -1,2 +1,0 @@
-class AutoOpNamespace:
-    ...
