--- iree/turbine/boo/bag/namespaces/auto_op/template_cache/README.md
@@ -1,1 +1,0 @@
-This dir should store a json with a mapping between template names and cached KernelTemplates (maybe cached torch.fx.GraphModules?)
