--- iree/turbine/boo/bag/namespaces/auto_op/templates.py
@@ -1,1 +1,0 @@
-# This should define an interface for seaching the template_cache for saved templates
