--- iree/turbine/boo/bag/registry.py
@@ -1,1 +1,0 @@
-# Setup a namespace registry. Each registered namespace should implement it's own cache manager.
